requirements.txt
web3==6.15.0
eth-utils==2.3.1
eth-abi==4.2.1
asyncio==3.4.3
websockets==12.0
python-dotenv==1.0.1
pydantic==2.6.4
SQLAlchemy==2.0.29
streamlit==1.32.2
pandas==2.2.1
scikit-learn==1.4.1.post1
torch==2.2.1
ollama==0.1.7
fastapi==0.110.0
uvicorn==0.29.0
loguru==0.7.2
pytest==8.1.1
